refactor(compiler-cli): remove dynamic require in ngcc to allow for bundling
ngcc currently dynamially loads the `Transformer` code. It does this
to avoid unnecessary parsing and loading of transformer-related code
if there is nothig to process (so-called noop case). Unfortunately
this dynamic require is not recognized by rollup (which we need to
use for shipping a CJS variant of the compiler-cli; and in the future
if we want to ship a fully-ESM compatible compiler-cli).

The ngcc code needs to use an async runtime import as it would work
in ES modules. This introduces async code into to the compililation
pipeline, breaking the `ngccMain` synchronous invocation feature.

To avoid this, we just move the dynamic require/async import to
the file top-level so that we do not break synchronous processing
which the CLI relies on. This has the downside of slowing-down
the noop case a little but I believe that should be mitigated
through bundling of ngcc anyway. In the future with full-ESM
we won't be able to get around this anyway (unless we remove the
sync variant of ngcc processing).