feat(core): add migration from synchronous fakeAsync to asynchronous variants
This migration is an experiment to convert `fakeAsync` usages to take
real asynchronous callbacks, to allow native async-await keywords to be
used without code transformation trickery.

The migration is limited in scope, in that functions that are migrated
to become asynchronous will not trigger their callers to await their
result (and possibly becoming async themselves, as a result).